* Generic Flexi   * Fixed: Exception when Send volume was mapped/changed and a FX or the Master track was selected. * Komplete Kontrol   * New: Added information about mute state due to soloed track(s)   * New: Added more selected track information to confirm to updated protocol   * New: Added scene navigation and play * OSC   * New: You can now configure the value range. So far, this was 128. Now you can increase the range to 1024 or 16384. Note that you also need to configure the widgets in your OSC client accordingly. * Code Cleanup.

+++ src/main/java/de/mossgrabers/controller/mcu/command/pitchbend/PitchbendVolumeCommand.java
@@ -37,7 +37,7 @@ public class PitchbendVolumeCommand extends AbstractPitchbendCommand<MCUControlS
     @Override
     public void onPitchbend (final int channel, final int data1, final int data2)
     {
-        final double value = Math.min (data2 * 127 + (double) data1, this.model.getValueChanger ().getUpperBound () - 1.0);
+        final int value = Math.min (data2 * 127 + data1, this.model.getValueChanger ().getUpperBound () - 1);
         if (channel == 8)
         {
             if (this.surface.isShiftPressed ())
@@ -76,7 +76,7 @@ public class PitchbendVolumeCommand extends AbstractPitchbendCommand<MCUControlS
             else if (modeManager.isActiveOrTempMode (Modes.MODE_SEND8))
                 track.getSendBank ().getItem (7).setValue (value);
             else if (modeManager.isActiveOrTempMode (Modes.MODE_DEVICE_PARAMS))
-                this.model.getCursorDevice ().getParameterBank ().getItem (extenderOffset + channel).setValue ((int) value);
+                this.model.getCursorDevice ().getParameterBank ().getItem (extenderOffset + channel).setValue (value);
             return;
         }
 
@@ -84,7 +84,7 @@ public class PitchbendVolumeCommand extends AbstractPitchbendCommand<MCUControlS
     }
 
 
-    private void handleTrack (final int index, final double value)
+    private void handleTrack (final int index, final int value)
     {
         final ITrack selectedTrack = this.model.getSelectedTrack ();
         switch (index)
